Switch from Palette to simple script
Palette has been great, but it's struggling to get by with some
highlight group names like the TreeSitter "@funcion.method" names
because of the magic it uses to turn bare words into highlight groups.

This sacrifices the cterm support which I may add back in, but it's so rare I
encounter a terminal that isn't 24-bit these days that I'm not sure it's
worth the burden anymore.

I considered trying to PR this against Palette or writing some sort of
migration script, but given how I'd already mixed logic and palette
together, I don't foresee it being easier than just macroing the change
to call `hi` instead.

Added the layer of indirection between `hi` and `vim_hi` as I may make
this export lua eventually if I want to make the theme more configurable
in some way for neovim users.